Work toward support of low-end devices.
I have a compute 50 device. To match exisitng code for compute 60 devices, I've
added blocks of comments containing new draft implementations for half-precision
fused-multiply-add, everywhere the __hfma functions are used.

I've also updated the include guards in exllama/cu_compat.cuh to reflect the
presence of a half2 atomicAdd function for compute 60 devices in CUDA versions
12.2 and newer.

The code now builds for my compute_50 device if the commented function blocks
are uncommented, although I lack the GPU RAM to run the testsuite on it.