Fix import of mapping from PEFT
The PEFT_TYPE_TO_MODEL_MAPPING variable was removed in PEFT PR
https://github.com/huggingface/peft/pull/2282. This leads to an error
when using AutoGPTQ with PEFT installed from main (and later starting
with PEFT v0.15.0).

This PR fixes this issue by using the PEFT_TYPE_TO_TUNER_MAPPING
variable instead. The fix is implemented in a way that the code should
continue to work with older PEFT versions while being compatible with
the new PR.